requirements.txt
google-api-python-client==2.19.0
watchdog==2.1.5
pydantic==1.8.2
